+++ Pinnacle/Core/UI/Builder/LabelButton.cs
@@ -10,7 +10,7 @@ using UnityEngine.UI;
 public sealed class LabelButton {
   public GameObject Container { get; private set; }
   public RectTransform RectTransform { get; private set; }
-  public TMP_Text Label { get; private set; }
+  public TextMeshProUGUI Label { get; private set; }
   public Button Button { get; private set; }
 
   public LabelButton(Transform parentTransform) {
@@ -39,8 +39,8 @@ public sealed class LabelButton {
     return container;
   }
 
-  static TMP_Text CreateLabel(Transform parentTransform) {
-    TMP_Text label = UIBuilder.CreateTMPLabel(parentTransform);
+  static TextMeshProUGUI CreateLabel(Transform parentTransform) {
+    TextMeshProUGUI label = UIBuilder.CreateTMPLabel(parentTransform);
 
     label
         .SetFontSize(16f)