+++ Pinnacle/Core/UI/Builder/ValueCell.cs
@@ -10,59 +10,68 @@ using UnityEngine;
 using UnityEngine.UI;
 
 public sealed class ValueCell {
-  public GameObject Cell { get; private set; }
+  public GameObject Container { get; private set; }
+  public RectTransform RectTransform { get; private set; }
+  public LayoutElement LayoutElement { get; private set; }
+
   public Image Background { get; private set; }
   public TMP_InputField InputField { get; private set; }
 
   public ValueCell(Transform parentTransform) {
-    Cell = CreateChildCell(parentTransform);
-    Background = Cell.Image();
+    Container = CreateContainer(parentTransform);
+    RectTransform = Container.GetComponent<RectTransform>();
+    LayoutElement = Container.GetComponent<LayoutElement>();
 
-    InputField = CreateChildInputField(Cell.transform);
+    Background = Container.GetComponent<Image>();
+
+    InputField = CreateValueInputField(Container.transform);
     InputField.SetTargetGraphic(Background);
   }
 
-  GameObject CreateChildCell(Transform parentTransform) {
-    GameObject cell = new("Cell", typeof(RectTransform));
-    cell.SetParent(parentTransform);
+  static GameObject CreateContainer(Transform parentTransform) {
+    GameObject container = new("Value", typeof(RectTransform));
+    container.transform.SetParent(parentTransform, worldPositionStays: false);
 
-    cell.AddComponent<Image>()
+    container.AddComponent<Image>()
         .SetType(Image.Type.Sliced)
-        .SetSprite(UIBuilder.CreateRoundedCornerSprite(64, 64, 8))
+        .SetSprite(UISpriteBuilder.CreateRoundedCornerSprite(64, 64, 8))
         .SetColor(new(0.5f, 0.5f, 0.5f, 0.5f));
 
-    cell.AddComponent<RectMask2D>();
-    cell.AddComponent<LayoutElement>();
+    container.AddComponent<RectMask2D>();
+    container.AddComponent<LayoutElement>();
 
-    return cell;
+    return container;
   }
 
-  TMP_InputField CreateChildInputField(Transform parentTransform) {
+  TMP_InputField CreateValueInputField(Transform parentTransform) {
     GameObject row = new("InputField", typeof(RectTransform));
     row.SetParent(parentTransform);
 
-    row.RectTransform()
+    row.GetComponent<RectTransform>()
         .SetAnchorMin(Vector2.zero)
         .SetAnchorMax(Vector2.one)
         .SetSizeDelta(Vector2.zero)
         .SetPosition(Vector2.zero);
 
-    TMP_Text label = UIBuilder.CreateTMPLabel(row.transform);
-    label.SetName("InputField.Text");
+    TextMeshProUGUI label = UIBuilder.CreateTMPLabel(row.transform);
 
-    label.richText = false;
-    label.alignment = TextAlignmentOptions.Left;
-    label.text = "InputField.Text";
+    label.SetName("Label")
+        .SetRichText(false)
+        .SetAlignment(TextAlignmentOptions.Left)
+        .SetText("Label");
 
     TMP_InputField inputField = parentTransform.gameObject.AddComponent<TMP_InputField>();
-    inputField.textComponent = label;
-    inputField.textViewport = row.GetComponent<RectTransform>();
-    inputField.transition = Selectable.Transition.ColorTint;
-    inputField.colors = InputFieldColorBlock.Value;
-    inputField.onFocusSelectAll = false;
-    inputField.SetNavigationMode(Navigation.Mode.None);
 
-    // row.AddComponent<DisableHighlightOnSelect>();
+    inputField
+        .SetTextComponent(label)
+        .SetTextViewport(row.GetComponent<RectTransform>())
+        .SetOnFocusSelectAll(false)
+        .SetDoubleClickDelay(0.2f);
+
+    inputField
+        .SetTransition(Selectable.Transition.ColorTint)
+        .SetColors(InputFieldColorBlock.Value)
+        .SetNavigationMode(Navigation.Mode.None);
 
     return inputField;
   }